test(api): move managed host fixture configs to model test support
Refactor shared DPU and managed-host test fixture config types into
`api-model::test_support`, and keep core-specific defaults in
`api-core::test_support::fixture_config`. Enable model test-support where
needed and update api-core/api-web tests to use the new fixture locations.
