+++ crates/api-core/src/test_support/mod.rs
@@ -18,6 +18,7 @@
 pub mod builder;
 pub mod default_config;
 pub mod endpoint_explorer;
+pub mod fixture_config;
 pub mod ib_fabric;
 pub mod ib_guid_pool;
 pub mod mac_address_pool;
