+++ crates/api-core/src/tests/instance_allocate.rs
@@ -23,13 +23,14 @@ use forge::forge_server::Forge;
 use ipnetwork::IpNetwork;
 use itertools::Itertools;
 use model::machine::{ManagedHostState, ManagedHostStateSnapshot};
+use model::test_support::ManagedHostConfig;
 use rpc::{Metadata, forge};
 
 use crate::cfg::file::{FnnConfig, FnnRoutingProfileConfig, PrefixFilterPolicyEntry};
+use crate::test_support::fixture_config::{FixtureDefault as _, ManagedHostConfigExt as _};
 use crate::test_support::mac_address_pool::HOST_NON_DPU_MAC_ADDRESS_POOL;
 use crate::tests::common;
 use crate::tests::common::api_fixtures;
-use crate::tests::common::api_fixtures::managed_host::ManagedHostConfig;
 use crate::tests::common::api_fixtures::network_segment::{
     FIXTURE_ADMIN_NETWORK_SEGMENT_GATEWAY, FIXTURE_HOST_INBAND_NETWORK_SEGMENT_GATEWAY,
     FIXTURE_HOST_INBAND_NETWORK_SEGMENT_GATEWAY_2, FIXTURE_TENANT_NETWORK_SEGMENT_GATEWAYS,
@@ -536,7 +537,7 @@ async fn test_zero_dpu_instance_allocation_auto_multi_segment(
             HOST_NON_DPU_MAC_ADDRESS_POOL.allocate(),
             HOST_NON_DPU_MAC_ADDRESS_POOL.allocate(),
         ],
-        ..Default::default()
+        ..ManagedHostConfig::default()
     };
 
     // Ingest zero DPU host with custom behavior in the finish callback...
@@ -712,7 +713,8 @@ async fn test_reject_single_dpu_instance_allocation_no_network_config(
     let env = create_test_env_for_instance_allocation(pool.clone(), None).await;
 
     // Create single DPU host
-    let single_dpu_host = api_fixtures::site_explorer::new_host(&env, Default::default()).await?;
+    let single_dpu_host =
+        api_fixtures::site_explorer::new_host(&env, ManagedHostConfig::default()).await?;
 
     // Create an instance on a host with DPUs, without specifying a network config, which is not allowed
     let result = crate::handlers::instance::allocate(
@@ -767,7 +769,8 @@ async fn test_reject_single_dpu_instance_allocation_host_inband_network_config(
     let env = create_test_env_for_instance_allocation(pool.clone(), None).await;
 
     // Create single DPU host
-    let single_dpu_host = api_fixtures::site_explorer::new_host(&env, Default::default()).await?;
+    let single_dpu_host =
+        api_fixtures::site_explorer::new_host(&env, ManagedHostConfig::default()).await?;
 
     let host_inband_segment =
         db::network_segment::find_by_name(env.pool.begin().await?.deref_mut(), "HOST_INBAND")
@@ -862,7 +865,7 @@ async fn test_reject_zero_dpu_instance_allocation_multiple_vpcs(
             HOST_NON_DPU_MAC_ADDRESS_POOL.allocate(),
             HOST_NON_DPU_MAC_ADDRESS_POOL.allocate(),
         ],
-        ..Default::default()
+        ..ManagedHostConfig::default()
     };
 
     // Ingest zero DPU host
@@ -990,7 +993,8 @@ async fn test_single_dpu_instance_allocation(
     let env = create_test_env_for_instance_allocation(pool.clone(), None).await;
 
     // Create single DPU host
-    let single_dpu_host = api_fixtures::site_explorer::new_host(&env, Default::default()).await?;
+    let single_dpu_host =
+        api_fixtures::site_explorer::new_host(&env, ManagedHostConfig::default()).await?;
 
     let tenant_segment =
         db::network_segment::find_by_name(env.pool.begin().await?.deref_mut(), "TENANT").await?;
@@ -1479,7 +1483,7 @@ async fn test_instance_allocation_rejects_auto_with_explicit_interfaces(
 async fn test_instance_allocation_rejects_auto_on_dpu_host(
     pool: sqlx::PgPool,
 ) -> Result<(), Box<dyn std::error::Error>> {
-    use crate::tests::common::api_fixtures::dpu::DpuConfig;
+    use model::test_support::DpuConfig;
 
     let env = create_test_env_for_instance_allocation(pool.clone(), None).await;
     // Default ManagedHostConfig has one DPU.
