+++ crates/api-core/src/tests/machine_dhcp.rs
@@ -20,7 +20,6 @@ use std::str::FromStr;
 
 use carbide_network::ip::IpAddressFamily;
 use carbide_uuid::machine::MachineInterfaceId;
-use common::api_fixtures::managed_host::ManagedHostConfig;
 use common::api_fixtures::network_segment::{
     FIXTURE_ADMIN_NETWORK_SEGMENT_GATEWAY, FIXTURE_HOST_INBAND_NETWORK_SEGMENT_GATEWAY,
     create_host_inband_network_segment, create_network_segment,
@@ -36,10 +35,12 @@ use itertools::Itertools;
 use mac_address::MacAddress;
 use model::machine_interface::InterfaceType;
 use model::network_segment::NetworkSegmentType;
+use model::test_support::ManagedHostConfig;
 use rpc::forge::ManagedHostNetworkConfigRequest;
 use rpc::forge::forge_server::Forge;
 
 use crate::DatabaseError;
+use crate::test_support::fixture_config::ManagedHostConfigExt as _;
 use crate::tests::common;
 use crate::tests::common::rpc_builder::DhcpDiscovery;
