+++ crates/api-core/src/tests/machine_find.rs
@@ -19,7 +19,6 @@ use std::net::IpAddr;
 use carbide_uuid::machine::{MACHINE_ID_PREFIX_LENGTH, MachineId, MachineType};
 use carbide_uuid::rack::RackId;
 use common::api_fixtures::dpu::create_dpu_machine;
-use common::api_fixtures::managed_host::ManagedHostConfig;
 use common::api_fixtures::{create_managed_host, create_test_env, site_explorer};
 use data_encoding::BASE32_DNSSEC;
 use db::ObjectFilter;
@@ -29,6 +28,7 @@ use model::expected_machine::ExpectedMachineData;
 use model::hardware_info::HardwareInfo;
 use model::machine::machine_id::host_id_from_dpu_hardware_info;
 use model::machine::machine_search_config::MachineSearchConfig;
+use model::test_support::ManagedHostConfig;
 use rpc::forge::forge_server::Forge;
 use rpc::forge::{
     AssociateMachinesWithInstanceTypeRequest, FindInstanceTypeIdsRequest, MachinesByIdsRequest,
@@ -36,6 +36,7 @@ use rpc::forge::{
 use sha2::{Digest, Sha256};
 use tonic::Request;
 
+use crate::test_support::fixture_config::{FixtureDefault as _, ManagedHostConfigExt as _};
 use crate::test_support::mac_address_pool::DPU_OOB_MAC_ADDRESS_POOL;
 use crate::tests::common;
 use crate::tests::common::api_fixtures::site_explorer::TestRackDbBuilder;
