+++ crates/api-core/src/tests/set_primary_interface.rs
@@ -20,12 +20,12 @@ use std::str::FromStr;
 use carbide_uuid::machine::MachineInterfaceId;
 use ipnetwork::IpNetwork;
 use model::network_segment::NetworkSegmentType;
+use model::test_support::{DpuConfig, ManagedHostConfig};
 use rpc::forge;
 use rpc::forge::forge_server::Forge;
 
+use crate::test_support::fixture_config::{FixtureDefault as _, ManagedHostConfigExt as _};
 use crate::tests::common::api_fixtures;
-use crate::tests::common::api_fixtures::dpu::DpuConfig;
-use crate::tests::common::api_fixtures::managed_host::ManagedHostConfig;
 use crate::tests::common::api_fixtures::network_segment::{
     FIXTURE_ADMIN_NETWORK_SEGMENT_GATEWAY, FIXTURE_HOST_INBAND_NETWORK_SEGMENT_GATEWAY,
     FIXTURE_UNDERLAY_NETWORK_SEGMENT_GATEWAY, create_admin_network_segment,
