+++ crates/api-model/Cargo.toml
@@ -82,6 +82,9 @@ version-compare = { workspace = true }
 tokio = { workspace = true }
 #these are alphabetized
 
+[features]
+test-support = []
+
 [build-dependencies]
 carbide-version = { path = "../version" }
